# back/api/openapi.yaml
openapi: 3.0.3
info:
  title: SeasonSchedule API
  description: API for managing team schedules and weekly event digests
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/subscribe:
    post:
      summary: Subscribe to weekly digest
      description: Add an email address to receive weekly schedule digests
      operationId: subscribe
      tags:
        - Subscriptions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
      responses:
        '201':
          description: Successfully subscribed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Successfully subscribed!
        '400':
          description: Invalid email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already subscribed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/events:
    get:
      summary: Get events
      description: Retrieve events, optionally filtered by week or team
      operationId: getEvents
      tags:
        - Events
      parameters:
        - name: week
          in: query
          description: ISO date (YYYY-MM-DD) of Sunday for the week
          required: false
          schema:
            type: string
            format: date
            example: "2026-04-19"
        - name: teamId
          in: query
          description: Filter by team UUID
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
    post:
      summary: Create event
      description: Create a new scheduled event (requires authentication)
      operationId: createEvent
      tags:
        - Events
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCreate'
      responses:
        '201':
          description: Event created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/events/{id}:
    put:
      summary: Update event
      description: Update an existing event (requires authentication)
      operationId: updateEvent
      tags:
        - Events
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCreate'
      responses:
        '200':
          description: Event updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete event
      description: Delete an event (requires authentication)
      operationId: deleteEvent
      tags:
        - Events
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Event deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/organizations/{orgId}/teams:
    get:
      summary: Get teams by organization
      description: Retrieve all teams within an organization
      operationId: getTeamsByOrganization
      tags:
        - Teams
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team'
  /api/organizations/{orgId}/events:
    get:
      summary: Get events by organization
      description: Retrieve all events for teams within an organization
      operationId: getEventsByOrganization
      tags:
        - Events
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
  /api/permissions:
    post:
      summary: Grant permission
      description: Grant a permission to a user for a resource (admin only)
      operationId: grantPermission
      tags:
        - Permissions
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PermissionGrant'
      responses:
        '201':
          description: Permission granted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Revoke permission
      description: Revoke a permission from a user (admin only)
      operationId: revokePermission
      tags:
        - Permissions
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PermissionGrant'
      responses:
        '200':
          description: Permission revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/users/{userId}/permissions:
    get:
      summary: Get user permissions
      description: Get all permissions for a user (admin only)
      operationId: getUserPermissions
      tags:
        - Permissions
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of permissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Permission'
        '403':
          description: Forbidden - admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Invalid credentials
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: "Team Practice"
        startDateTime:
          type: string
          format: date-time
          example: "2026-04-21T10:00:00Z"
        endDateTime:
          type: string
          format: date-time
          example: "2026-04-21T12:00:00Z"
        location:
          type: string
          example: "Main Field"
        teamId:
          type: string
          format: uuid
        teamName:
          type: string
          example: "Varsity Basketball"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    EventCreate:
      type: object
      required:
        - title
        - startDateTime
        - teamId
      properties:
        title:
          type: string
          example: "Team Practice"
        startDateTime:
          type: string
          format: date-time
          example: "2026-04-21T10:00:00Z"
        endDateTime:
          type: string
          format: date-time
          example: "2026-04-21T12:00:00Z"
        location:
          type: string
          example: "Main Field"
        teamId:
          type: string
          format: uuid
    Team:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        name:
          type: string
          example: "Varsity Basketball"
        organizationName:
          type: string
          example: "Sports Club"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Permission:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        resourceType:
          type: string
          enum: [organization, team]
        resourceId:
          type: string
          format: uuid
        permission:
          type: string
          enum: [read, write, admin]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PermissionGrant:
      type: object
      required:
        - userId
        - resourceType
        - resourceId
        - permission
      properties:
        userId:
          type: string
          format: uuid
        resourceType:
          type: string
          enum: [organization, team]
        resourceId:
          type: string
          format: uuid
        permission:
          type: string
          enum: [read, write, admin]